r5817/sofa-dev : FIX : Removes some Win32 warnings.
 //partial sync (2/3 changes)//

+++ modules/sofa/component/visualmodel/FlowVisualModel.inl
@@ -195,7 +195,7 @@ void FlowVisualModel<DataTypes>::initVisual()
         tetraShellPerTriangleVertex.resize(nbPoints);
         isPointInTetra.resize(nbPoints);
         tetraSize.resize(nbTetrahedra);
-        std::fill(tetraSize.begin(), tetraSize.end(), 0.0);
+        std::fill(tetraSize.begin(), tetraSize.end(), (float)0.0);
         std::fill(isPointInTetra.begin(), isPointInTetra.end(), true);
 
         //Store some data
